requirements.txt
MetaTrader5>=5.0.4682
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0 